// src/api/generated/schemas/chatListItemDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Chat Backend API
 * OpenAPI spec version: 1.0
 */
import type { ChatListItemDtoType } from './chatListItemDtoType';
import type { ChatLastMessageDto } from './chatLastMessageDto';

export interface ChatListItemDto {
  id: string;
  type: ChatListItemDtoType;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  lastMessage?: ChatLastMessageDto;
  unreadCount: number;
}